The doctor table has been changed
