+++ package.json
@@ -12,10 +12,6 @@
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.4.0",
     "bootstrap": "^5.3.1",
-    "dotenv": "^16.3.1",
-    "i": "^0.3.7",
-    "mdb-ui-kit": "^6.4.1",
-    "npm": "^9.8.1",
     "react": "^18.2.0",
     "react-autosuggest": "^10.1.0",
     "react-bootstrap": "^2.8.0",
