+++ src/Components/Doctor/DoctorList.jsx
@@ -1,12 +1,11 @@
 import React, { useContext } from "react";
 import { GetTheAppContext } from "../../Context/AppContext";
-import { NewTable } from "./NewTable";
-// import { TablaGeneric } from "./table";
+import { TableDoctor } from "./TableDoctor";
+
 const DoctorList = () => {
   const { dataGetAllDoctors } = useContext(GetTheAppContext);
-
-  return <NewTable dataTable={dataGetAllDoctors || []} />;
-  // return <TablaGeneric title="MÉDICOS" data={dataGetAllDoctors || []} />;
+  return (
+    <TableDoctor dataTable={dataGetAllDoctors || []} />
+  );
 };
-
 export default DoctorList;
