+++ src/Css/App.css
@@ -1,15 +1,42 @@
 * {
-    padding: 0;
-    margin: 0;
-    box-sizing: border-box;
-  }
+  padding: 0;
+  margin: 0;
+  box-sizing: border-box;
+}
 
-  #idBarNavContent{
-    z-index: 200;
+#idBarNavContent {
+  z-index: 200;
+}
+
+#DivinputSearch {
+  border: none;
+  border-bottom: solid 0.5px #8c8c8e;
+  box-shadow: inset;
+}
+
+#inputSearch {
+  border: none;
+  font-size: 18px;
+  line-height: 1.5;
+  margin: 10px 0;
+  font-weight: 300;
+
+  vertical-align: middle;
+  border-color: #ddd;
+  font-weight: 300;
+  font-size: 16px;
+  color: #3c4858;
+}
+
+#inputSearch:focus {
+  border: none;
+  box-shadow: none;
+}
+
+/* TODO: This section is to form all crud */
+
+@media (max-width: 900px) {
+  #disableCell {
+    display: none;
   }
-  
-  @media (max-width: 900px) {
-    #disableCell {
-      display: none;
-    }
-  }
+}
