+++ src/Css/DoctorCss.css
@@ -1,7 +1,33 @@
+/* TODO: This is my own css code */
 #btnTable {
   height: 15px;
   padding-bottom: 5px;
 }
+
+#btnAdd {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  background-color: #0d6efd;
+  border: solid 1px #bcbcbc;
+  height: 30px;
+  width: 60px;
+}
+
+#iconoClear {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  background-color: #0d6efd;
+  border: solid 1px #bcbcbc;
+  height: 40px;
+  width: 45px;
+  margin-top: 35px;
+}
+
+/* TODO: ------------------- */
 
 .container h3 {
   font-size: 25px;
@@ -11,6 +37,13 @@
   color: #3c4858;
 }
 .container h4 {
+  font-size: 18px;
+  line-height: 1.5;
+  margin: 10px 0;
+  font-weight: 300;
+  color: #3c4858;
+}
+.container option {
   font-size: 18px;
   line-height: 1.5;
   margin: 10px 0;
@@ -44,13 +77,7 @@
   border-top-left-radius: 0 !important;
   border-bottom-left-radius: 0 !important;
 }
-.btn {
-  padding: 12px 30px !important;
-  margin: 5px 1px;
-  font-size: 12px !important;
-  box-shadow: 0 2px 2px 0 hsla(0, 0%, 60%, 0.14),
-    0 3px 1px -2px hsla(0, 0%, 60%, 0.2), 0 1px 5px 0 hsla(0, 0%, 60%, 0.12);
-}
+
 .btn .material-icons {
   position: relative;
   display: inline-block;
@@ -71,7 +98,6 @@
 }
 
 .btn.btn-just-icon {
-  font-size: 24px;
   height: 41px;
   min-width: 41px;
   width: 41px;
@@ -83,86 +109,7 @@
 
 .btn.btn-just-icon.btn-round {
   border-radius: 50% !important;
-}
-
-.btn.btn-link {
-  background: transparent;
-  box-shadow: none;
-  color: #999;
-}
-
-.btn.btn-info {
-  color: #fff !important;
-  background-color: #00bcd4 !important;
-  border-color: #00bcd4;
-  box-shadow: 0 2px 2px 0 rgba(0, 188, 212, 0.14),
-    0 3px 1px -2px rgba(0, 188, 212, 0.2), 0 1px 5px 0 rgba(0, 188, 212, 0.12) !important;
-}
-
-.btn.btn-info:hover {
-  box-shadow: 0 14px 26px -12px rgba(0, 188, 212, 0.42),
-    0 4px 23px 0 rgba(0, 0, 0, 0.12), 0 8px 10px -5px rgba(0, 188, 212, 0.2) !important;
-  background: #00aec5 !important;
-}
-
-.btn.btn-info.btn-link {
-  background-color: transparent !important;
-  color: #00bcd4 !important;
-  box-shadow: none !important;
-}
-.btn.btn-success {
-  color: #fff !important;
-  background-color: #4caf50 !important;
-  border-color: #4caf50;
-  box-shadow: 0 2px 2px 0 rgba(76, 175, 80, 0.14),
-    0 3px 1px -2px rgba(76, 175, 80, 0.2), 0 1px 5px 0 rgba(76, 175, 80, 0.12) !important;
-}
-
-.btn.btn-success:hover {
-  box-shadow: 0 14px 26px -12px rgba(76, 175, 80, 0.42),
-    0 4px 23px 0 rgba(0, 0, 0, 0.12), 0 8px 10px -5px rgba(76, 175, 80, 0.2) !important;
-  background: #47a44b !important;
-}
-.btn.btn-success.btn-link {
-  background-color: transparent !important;
-  color: #4caf50 !important;
-  box-shadow: none !important;
-}
-
-.btn.btn-danger {
-  color: #fff !important;
-  background-color: #f44336 !important;
-  border-color: #f44336;
-  box-shadow: 0 2px 2px 0 rgba(244, 67, 54, 0.14),
-    0 3px 1px -2px rgba(244, 67, 54, 0.2), 0 1px 5px 0 rgba(244, 67, 54, 0.12) !important;
-}
-
-.btn.btn-danger:hover {
-  box-shadow: 0 14px 26px -12px rgba(244, 67, 54, 0.42),
-    0 4px 23px 0 rgba(0, 0, 0, 0.12), 0 8px 10px -5px rgba(244, 67, 54, 0.2) !important;
-  background-color: #f33527 !important;
-}
-
-.btn.btn-danger.btn-link {
-  background-color: transparent !important;
-  color: #f44336 !important;
-  box-shadow: none !important;
-}
-.btn.btn-just-icon .material-icons {
-  margin-top: 0;
-  position: absolute;
-  width: 100%;
-  transform: none;
-  left: 0;
-  top: 0;
-  height: 100%;
-  line-height: 41px;
-  font-size: 20px;
-}
-
-.btn.btn-just-icon.btn-sm .material-icons {
-  font-size: 17px;
-  line-height: 29px;
+  margin-right: 1px;
 }
 
 .table {
@@ -183,10 +130,9 @@
 }
 
 .table > tbody > tr > td {
-  /* padding: 12px 8px; */
   vertical-align: middle;
   border-color: #ddd;
   font-weight: 300;
-  font-size: 14px;
+  font-size: 16px;
   color: #3c4858;
 }
