+++ src/Css/navBar.css
@@ -1,17 +1,3 @@
-body {
-  position: relative;
-  overflow-x: hidden;
-  /* background-color: #cfd8dc; */
-}
-body,
-html {
-  height: 100%;
-}
-.nav .open > a,
-.nav .open > a:hover,
-.nav .open > a:focus {
-  background-color: transparent;
-}
 
 .my-custom-margin {
   margin-left: 4rem; /* Cambia este valor según tus necesidades */
